--- libs/components/manifest/src/generator/validation.ts
@@ -1,19 +1,0 @@
-import { type PackagesMap } from './get-public-api';
-
-export function validateDocsIds(packages: PackagesMap): string[] {
-  const errors: string[] = [];
-  const ids: string[] = [];
-
-  for (const [, definitions] of packages) {
-    for (const definition of definitions) {
-      if (ids.includes(definition.docsId)) {
-        errors.push(`Duplicate @docsId encountered: ${definition.docsId}`);
-        continue;
-      }
-
-      ids.push(definition.docsId);
-    }
-  }
-
-  return errors;
-}
